chore: remove "extracted" directory
- move VirtualizedTree back into components folder
- allow components to have a "util" sidecar
