--- src/Frontend/extracted/VirtualisedTree/Icons.tsx
@@ -1,29 +1,0 @@
-// SPDX-FileCopyrightText: Meta Platforms, Inc. and its affiliates
-// SPDX-FileCopyrightText: TNG Technology Consulting GmbH <https://www.tngtech.com>
-//
-// SPDX-License-Identifier: Apache-2.0
-import { SxProps } from '@mui/material';
-import MuiBox from '@mui/material/Box';
-import { ReactElement } from 'react';
-
-interface NodeIconProps {
-  sx?: SxProps;
-  ariaLabel: string;
-  onClick: () => void;
-  icon: ReactElement;
-}
-
-export function NodeIcon(props: NodeIconProps): ReactElement {
-  return (
-    <MuiBox
-      sx={props.sx}
-      onClick={(event): void => {
-        event.stopPropagation();
-        props.onClick();
-      }}
-      aria-label={props.ariaLabel}
-    >
-      {props.icon}
-    </MuiBox>
-  );
-}
